UPD: Moved TCP/UDP checksum calc
